backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.0.0
requests>=2.32.0
pywebview>=5.0
pystray>=0.19.5
pillow>=10.0.0
pyinstaller>=6.0.0
